+++ packages/cli/examples/get_akash_vesting_account.ts
@@ -1,6 +1,6 @@
 import { QueryClient, setupAuthExtension } from "@cosmjs/stargate";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
-import { Any } from "@quadrateorg/quadjs-types/google/protobuf/any";
+import { Any } from "cosmjs-types/google/protobuf/any";
 
 // https://github.com/ovrclk/net/blob/24ddbb427/mainnet/rpc-nodes.txt
 const tmClient = await Tendermint34Client.connect("http://rpc.akash.forbole.com:80");