+++ packages/cosmwasm-stargate/src/cosmwasmclient.searchtx.spec.ts
@@ -18,7 +18,7 @@ import {
   isMsgSendEncodeObject,
 } from "@cosmjs/stargate";
 import { assert, sleep } from "@cosmjs/utils";
-import { TxRaw } from "@quadrateorg/quadjs-types/cosmos/tx/v1beta1/tx";
+import { TxRaw } from "cosmjs-types/cosmos/tx/v1beta1/tx";
 
 import { CosmWasmClient } from "./cosmwasmclient";
 import {