+++ packages/cosmwasm-stargate/src/modules/wasm/queries.spec.ts
@@ -12,8 +12,8 @@ import {
   StdFee,
 } from "@cosmjs/stargate";
 import { assert, assertDefined } from "@cosmjs/utils";
-import { MsgExecuteContract, MsgInstantiateContract, MsgStoreCode } from "@quadrateorg/quadjs-types/cosmwasm/wasm/v1/tx";
-import { AbsoluteTxPosition, ContractCodeHistoryOperationType } from "@quadrateorg/quadjs-types/cosmwasm/wasm/v1/types";
+import { MsgExecuteContract, MsgInstantiateContract, MsgStoreCode } from "cosmjs-types/cosmwasm/wasm/v1/tx";
+import { AbsoluteTxPosition, ContractCodeHistoryOperationType } from "cosmjs-types/cosmwasm/wasm/v1/types";
 import Long from "long";
 
 import { SigningCosmWasmClient } from "../../signingcosmwasmclient";