+++ packages/stargate/src/accounts.ts
@@ -2,16 +2,15 @@ import { Pubkey } from "@cosmjs/amino";
 import { Uint64 } from "@cosmjs/math";
 import { decodePubkey } from "@cosmjs/proto-signing";
 import { assert } from "@cosmjs/utils";
-import { BaseAccount, ModuleAccount } from "@quadrateorg/quadjs-types/cosmos/auth/v1beta1/auth";
+import { BaseAccount, ModuleAccount } from "cosmjs-types/cosmos/auth/v1beta1/auth";
 import {
   BaseVestingAccount,
   ContinuousVestingAccount,
   DelayedVestingAccount,
   PeriodicVestingAccount,
-} from "@quadrateorg/quadjs-types/cosmos/vesting/v1beta1/vesting";
-import { Any } from "@quadrateorg/quadjs-types/google/protobuf/any";
+} from "cosmjs-types/cosmos/vesting/v1beta1/vesting";
+import { Any } from "cosmjs-types/google/protobuf/any";
 import Long from "long";
-import { EthAccount } from "@quadrateorg/quadjs-types/ethermint/v1beta1/auth";
 
 export interface Account {
   /** Bech32 account address */
@@ -27,18 +26,6 @@ function uint64FromProto(input: number | Long): Uint64 {
 
 function accountFromBaseAccount(input: BaseAccount): Account {
   const { address, pubKey, accountNumber, sequence } = input;
-  const pubkey = pubKey ? decodePubkey(pubKey) : null;
-  return {
-    address: address,
-    pubkey: pubkey,
-    accountNumber: uint64FromProto(accountNumber).toNumber(),
-    sequence: uint64FromProto(sequence).toNumber(),
-  };
-}
-
-function accountFromEthAccount(input: EthAccount): Account {
-  const { base_account, code_hash } = input;
-  const { address, pubKey, accountNumber, sequence } = base_account;
   const pubkey = pubKey ? decodePubkey(pubKey) : null;
   return {
     address: address,
@@ -64,10 +51,6 @@ export function accountFromAny(input: Any): Account {
 
   switch (typeUrl) {
     // auth
-
-    case "/ethermint.types.v1.EthAccount": {
-      return accountFromEthAccount(EthAccount.decode(value));
-    }
 
     case "/cosmos.auth.v1beta1.BaseAccount":
       return accountFromBaseAccount(BaseAccount.decode(value));