+++ packages/stargate/src/modules/auth/queries.spec.ts
@@ -2,8 +2,8 @@
 import { encodePubkey } from "@cosmjs/proto-signing";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { assert } from "@cosmjs/utils";
-import { BaseAccount } from "@quadrateorg/quadjs-types/cosmos/auth/v1beta1/auth";
-import { Any } from "@quadrateorg/quadjs-types/google/protobuf/any";
+import { BaseAccount } from "cosmjs-types/cosmos/auth/v1beta1/auth";
+import { Any } from "cosmjs-types/google/protobuf/any";
 import Long from "long";
 
 import { QueryClient } from "../../queryclient";