+++ packages/stargate/src/modules/tx/queries.spec.ts
@@ -1,7 +1,7 @@
 import { coin, coins, DirectSecp256k1HdWallet, Registry } from "@cosmjs/proto-signing";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { assertDefined, sleep } from "@cosmjs/utils";
-import { MsgDelegate } from "@quadrateorg/quadjs-types/cosmos/staking/v1beta1/tx";
+import { MsgDelegate } from "cosmjs-types/cosmos/staking/v1beta1/tx";
 import Long from "long";
 
 import { QueryClient } from "../../queryclient";