+++ packages/stargate/src/modules/vesting/messages.ts
@@ -1,5 +1,5 @@
 import { GeneratedType } from "@cosmjs/proto-signing";
-import { MsgCreateVestingAccount } from "@quadrateorg/quadjs-types/cosmos/vesting/v1beta1/tx";
+import { MsgCreateVestingAccount } from "cosmjs-types/cosmos/vesting/v1beta1/tx";
 
 export const vestingTypes: ReadonlyArray<[string, GeneratedType]> = [
   ["/cosmos.vesting.v1beta1.MsgCreateVestingAccount", MsgCreateVestingAccount],