+++ packages/stargate/src/queryclient/queryclient.spec.ts
@@ -4,13 +4,13 @@ import { toAscii } from "@cosmjs/encoding";
 import { DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
 import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { assert } from "@cosmjs/utils";
-import { Metadata } from "@quadrateorg/quadjs-types/cosmos/bank/v1beta1/bank";
+import { Metadata } from "cosmjs-types/cosmos/bank/v1beta1/bank";
 import {
   QueryAllBalancesRequest,
   QueryAllBalancesResponse,
   QueryBalanceRequest,
   QueryBalanceResponse,
-} from "@quadrateorg/quadjs-types/cosmos/bank/v1beta1/query";
+} from "cosmjs-types/cosmos/bank/v1beta1/query";
 
 import { SigningStargateClient } from "../signingstargateclient";
 import {